+++ src/adapter/cloudflare-workers/serve-static.ts
@@ -1,7 +1,7 @@
 // @denoify-ignore
 import type { KVNamespace } from '@cloudflare/workers-types'
 import type { MiddlewareHandler } from '../../types'
-import { getContentFromKVAsset } from '../../utils/cloudflare'
+import { getContentFromKVAsset } from './cloudflare'
 import { getFilePath } from '../../utils/filepath'
 import { getMimeType } from '../../utils/mime'
